+++ bff/lkshmatch/.env.example
@@ -1,0 +1,4 @@
+export BOT_TOKEN="bot-token"
+export CREDENTIALS_FILE = "Путь до файла с ключом для сервисного аккаунта google cloud"
+export SERVICE_ACCOUNT_NAME = "Почта сервисного аккаунта google cloud"
+export JWT_SECRET_KEY = "jwt-secret-key"
