+++ bff/lkshmatch/config.py
@@ -6,5 +6,5 @@ settings = Dynaconf(
     environments=True,
     load_dotenv=True,
     default_env="prod",
-    env="test",
+    env="test"
 )
