+++ bff/requirements.txt
@@ -6,3 +6,8 @@ pydantic
 dynaconf
 python-jose
 pytelegrambotapi
+dishka
+httplib2
+google-api-python-client
+oauth2client
+aiohttp
